* Custom Reflection utilities written * A lot of bugs fixed  - WrappedPacketInFlying's subclasses, the isPosition and isLook method were false sometimes  - WrappedPacketInBlockPlace fixed  - Many optimizations
 * All serverbound wrappers are STABLE, the Clientbound wrappers haven't been tested
 * Massive code cleanup

+++ src/main/java/io/github/retrooper/packetevents/enums/Direction.java
@@ -9,7 +9,7 @@ public enum Direction {
     SOUTH_SOUTH_EAST, SOUTH_SOUTH_WEST,
     WEST_SOUTH_WEST, SELF, NULL;
 
-    public static Direction get(final byte i) {
+    public static Direction get(final int i) {
         return values()[i];
     }
 }
