+++ src/main/java/io/github/retrooper/packetevents/enums/minecraft/EntityUseAction.java
@@ -3,7 +3,7 @@ package io.github.retrooper.packetevents.enums.minecraft;
 public enum EntityUseAction {
     INTERACT, INTERACT_AT, ATTACK, INVALID;
 
-    public static EntityUseAction get(final byte i) {
+    public static EntityUseAction get(final int i) {
         return values()[i];
     }
 }
